set Measurement as generic data attribute in kaleido parser
